Move stack​ path handling into metadata
