+++ lib/cuffsert/main.rb
@@ -13,27 +13,10 @@ require 'uri'
 # - Stop using file: that we anyway need to special-case in cfarguments
 # - default value for meta.metadata when stack_path is local file
 # - selector and metadata are mandatory and need guards accordingly
-# - validate_and_urlify belongs in metadata.rb
 # - execute should use helpers and not know details of statuses
 # - update 'abort' should delete cheangeset and emit the result
 
 module CuffSert
-  def self.validate_and_urlify(stack_path)
-    if stack_path =~ /^[A-Za-z0-9]+:/
-      stack_uri = URI.parse(stack_path)
-    else
-      normalized = File.expand_path(stack_path)
-      unless File.exist?(normalized)
-        raise "Local file #{normalized} does not exist"
-      end
-      stack_uri = URI.join('file:///', normalized)
-    end
-    unless ['s3', 'file'].include?(stack_uri.scheme)
-      raise "Uri #{stack_uri.scheme} is not supported"
-    end
-    stack_uri
-  end
-
   def self.create_stack(client, meta, confirm_create)
     cfargs = CuffSert.as_create_stack_args(meta)
     Rx::Observable.concat(
@@ -117,8 +100,6 @@ module CuffSert
     cli_args = CuffSert.parse_cli_args(argv)
     CuffSert.validate_cli_args(cli_args)
     meta = CuffSert.build_meta(cli_args)
-    stack_path = cli_args[:stack_path][0]
-    meta.stack_uri = CuffSert.validate_and_urlify(stack_path)
     events = CuffSert.execute(meta, CuffSert.method(:confirmation),
       force_replace: cli_args[:force_replace])
     renderer = CuffSert.make_renderer(cli_args)
